Move pthread linkopts declaration to the cc_binary
Before, we had a different treatement of linkopts for aarch64 and
x86_64. So x86_64 would declare pthread linking in the bazelrc and the
aarch64 would declare it in the toolchain file.

This is an attempt to harmonize both approaches. And since we can't
change the bazel toolchain definition for x86_64, the proposed approach
changes the aarch64 toolchain definition instead.

This change then adds the pthread linking in the linkopts, making it
optional.