Restructure the setup and build doc
I have moved the information about collect_deps in the setup part, since
it fits better there. I have also removed the duplicated build steps
when using the build_env and not. The build instructions are the same
for both cases. The only difference being the act of entering the build
environment.