Add pthread to CMake cache_entries
There are two executables that get built by the BSW modules, the log
daemon and the executable manager. We give them both instructions to
link pthread statically.

This is not normally needed, but with CMAKE_TRY_COMPILE_TARGET_TYPE set
to STATIC_LIBRARY, this needs to be instructed at compile time.