Harmonize TARGET_TYPE to STATIC
I do not know exactly why using this option affects the linking of
pthread, but it seems it does, so we're linking pthread manually where
it's needed. Furthermore, we make the approach uniform across all the
targets so that we don't have any mismatches that might be harder to
debug in the future.