Restructure bazelrc to reveal more commonalities
First, this change removes the outdated comment about pthreads. This is
no longer applicable in the latest master.

Furthermore, we pull out the commonality of experimental_scale_timeouts
for both configurations.

We also put the build_target configs side to side to show what's
configured for each platform.

Finally we group crosstool_top and host_crosstool_top since
host_crosstool_top is mainly needed as a consequence of setting
corsstool_top in the first place.