Print code gen warnings and errors in stdout
If xmllint is detected, then use it to parse the GeneratorReport.xml and
print the non-fatal errors and warnings on stdout. If it is not
installed, then simply check for errors and warnings, and print a link
to the log file in stdout.

xmllint seems to be present on my machine without me installing it, but
we might have to install it in the docker container.