Fix building from the hermetized toolchain
Without the first line in this change, only the contents of usr/lib were
getting added to the sandbox, which meant things like toolchain binaries
such as the as, ar, ld tool were not properly included. This adds
everything under usr.

As for the second line in this change, it's a future proofing for when
we add support for aarch64 toolchains for qnx