Hardcode pthread in the BSW Linux config
It seems that find_package(Threads) does not work well inside bazel.
It's not fully clear why this is happening and the CMake code for this
is really hard to understand. Instead, let's hardcode this for builds
targetting Linux. This should not add any harm. On bazel-only builds,
this is handled using toolchain features.