Implement socket and pthread features for aarch64
Implement this new way to handle pthread and socket dependencies to
conditionally when switching for different toolchains and OSes.

This is useful, for example, because on QNX you need to link socket
libraries but don't need to link pthread. On linux, this is different:
you don't need to link socket  but you need to link pthread.