Add x86_64_linux pthread linkopt in the bazelrc
We are adding this as a temporary workaround until we write the
x86_64_linux toolchain as its own file instead of using the one from
bazel tools. It is necessary to write an own file because now we are
using the bazel toolchain features to indicate when socket and pthreads
are necessary.