Add qnx x86_64 toolchain settings
The current implementation builds correctly, but the flags are not very
clean. This is mainly because our old gcc toolchain files were also very
messy.

With this configuration, if the QNX_HOST and QNX_TARGET environment
variables are set, then they will be used. If they are not set, then
it will try to assume a default QNX install path of /opt/qnx/qnx700.