Add explicit compiler spec files
This enables us to have more fine-grained control over the default flags
of the compiler. It is currently necessary because it seems like by
default, the spec file points to the native `as` binary instead of the
platform specific one embedded into the toolchain. In qcc this also
appears to be worked-around with the spec file.