Fix qnx toolchain invoke_as
Without this change, the toolchain ends up using the native `as` binary
installed in the system, and not the `as` binary embedded in the
toolchain. This is obviously not very hermetic.

Luckily building for QNX x86_64 worked without this fix because the `as`
binary on my laptop builds for x86_64.

But this becomes a problem for aarch64, and this is where I started
noticing the issue.

qcc is taking a similar approach in its conf files:

  as=$(QNX_HOST)/usr/bin/x86_64-pc-nto-qnx7.0.0-as