Make circular dependency workaround use cpu flag
Since now we have this tag available, we can use it instead of the
build_target one. It is much better suitable.

It's still not perfect, because you can build linux binaries
and maybe link them to qnx binaries, but circular dependency workarounds
shouldn't be needed in the first place...