Export out_include_dirs to dependents and to FLAGS
We need the out_include_dirs to be exported to the FLAGS variables so we
can link collectd plugins against BSW libraries (for example ara::log
for the write_dlt plugin).

For this we need to export this using the Bazel provider and then
consume this when the FLAGS variables are constructed.